CS /DOT-js MSTR CHAIN INTERGRATIONas
polka Dot JS application integrated with original blockchain data sets. Larger data sets and multiplied computational elements would seem to make block times longer, that is not the case at 5.5-6.4 seconds on average.  It will become of great use if utilized appropriately.  DOT token is and will be a standard within the industry as it scales well and plays well within the system as a assistant to many processes at a low rate for transport and auxiliary nature needed for expedited processing.